requirements.txt
six>=1.12.0
wheel>=0.33.1
setuptools>=40.8.0

pyyaml>=5.3.1,<6.0.0
jsonschema >=3.2.0,<4.0.0
ijson >=3.1.4,<4.0.0
lxml >= 4.6.3
xmltodict >= 0.12.0

requests>=2.24.0,<3.0.0
httplib2>=0.20.4, < 0.21.0
websockets>=8.0.0,<9.0.0

kaitaistruct==0.9
aiohttp>=3.8.1,<4.0.0

numpy>=1.19.0,<1.24.0
pillow>=8.1.0,<9.0.0

pyzmq>=22.0.0,<23.0.0
protobuf==3.15.5

wrapt>=1.12.1,<2.0.0

pymongo>=4.0.1

importlib_resources; python_version < "3.9"